@@ -47,7 +47,7 @@
   },
   "devDependencies": {
     "@antfu/eslint-config": "^0.43.1",
-    "@percy/cli": "^1.27.0",
+    "@percy/cli": "^1.30.0",
     "@percy/cypress": "^3.1.2",
     "@semantic-release/changelog": "^6.0.3",
     "@semantic-release/git": "^10.0.1",
@@ -58,10 +58,11 @@
     "@storybook/testing-library": "^0.2.0",
     "@storybook/vue3": "^7.1.1",
     "@storybook/vue3-vite": "^7.1.1",
-    "@types/node": "^20.4.9",
-    "@vitejs/plugin-vue": "^5.0.5",
-    "@vue/tsconfig": "^0.1.3",
-    "cypress": "^13.0.0",
+    "@tsconfig/node18": "^18.2.4",
+    "@types/node": "^18.19.55",
+    "@vitejs/plugin-vue": "^5.1.4",
+    "@vue/tsconfig": "^0.5.1",
+    "cypress": "^13.15.0",
     "eslint": "^8.49.0",
     "eslint-plugin-cypress": "^2.13.3",
     "eslint-plugin-storybook": "^0.6.13",
@@ -69,28 +70,28 @@
     "prettier": "^2.8.8",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "sass": "^1.64.2",
+    "sass": "^1.79.5",
     "storybook": "^7.4.1",
-    "typescript": "^5.2.2",
+    "typescript": "^5.6.3",
     "ucla-library-design-tokens": "^5.23.0",
     "video.js": "^8.5.2",
-    "vite": "^5.3.1",
-    "vite-svg-loader": "^4.0.0",
-    "vue": "^3.4.29",
-    "vue-router": "^4.3.3",
-    "vue-tsc": "^2.0.24"
+    "vite": "^5.4.8",
+    "vite-svg-loader": "^5.1.0",
+    "vue": "^3.5.12",
+    "vue-router": "^4.4.5",
+    "vue-tsc": "^2.1.6"
   },
   "dependencies": {
     "@vuepic/vue-datepicker": "^8.2.0",
-    "@vueuse/components": "^10.11.0",
-    "@vueuse/core": "^10.11.0",
+    "@vueuse/components": "^11.1.0",
+    "@vueuse/core": "^11.1.0",
     "add-to-calendar-button": "^2.6.18",
     "date-fns": "^2.30.0",
     "lodash": "^4.17.21",
-    "pinia": "^2.1.6",
+    "pinia": "^2.2.4",
     "vue3-carousel": "^0.3.1"
   },
   "peerDependencies": {
-    "vue": "^3.4.29"
+    "vue": "^3.5.12"
   }
 }
